Replace 4 * SIGSTKSZ with 32768
Recent libc does not provide SIGSTKSZ (nor MINSIGSTKSZ, it
would appear) as a constant anymore. Instead, it is
retrieved at runtime via sysconf. It is thus unusable as
a static array size specification. Given that we were
already using the fairly arbitrary 4 * SIGSTKSZ, just go
ahead and take 32KB. We can't check for the definition of
the macro, since SIGSTKSZ is implemented as a macro to
sysconf(). 32KB really ought be both sufficient to handle
a signal stack, and available, on all but the most
exotic architectures. This is the value to which 4 * SIGSTKSZ
evaluated on both Debian and Fedora. It's better to have the
static altstack area than have to allocate one at runtime,
it eliminates the need to handle a bogon return from sysconf,
and should it be too extravagant on some system, the binary
will fail at startup rather than mysteriously in media res.