Fix build with -Wunused-but-set-variable
Some compilers (hi clang 14) warn about variables which are written to
but not used -- and that's what happens when doctest is disabled by a
compiler macro in the `disabled` tests.

Fix by simply returning the result from that function. That's less
invasive than printing to stdout which would break the testsuite.