Updated platform checks, esp Apple
The existing platform checks assumed that <TargetConditionals.h> was
already included prior to doctest.h. If it wasn't, DOCTEST_PLATFORM_MAC
wouldn't get set because doctest was checking for macros defined in
that header.

This caused issues in cross-platform code being run/tested on an
Apple platform; I first noticed that breakpoints on CHECK didn't work.

Since the comment above said "this is how Catch does it", I went to
a recent version of Catch and copied/adapted its platform check,
which does work properly. I changed it a bit because there's no real
need to distinguish between macOS and iOS or other Apple platforms
(AFAIK) since they work the same under the hood; DOCTEST_PLATFORM_MAC
will be set on any Apple platform. If there's a need to specialize,
there are several other macros in TargetConditionals.h that can be
checked.